+++ vscode-lox/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add support for `argv` global variable when extra features are enabled.
+
 ### Fixed
 
 - Report write-only property error when getter of different staticness exists.